sim: change host_uart_puts & gets return type
from ssize to int, for windows don't support ssize

+++ arch/sim/src/sim/posix/sim_hostuart.c
@@ -134,7 +134,7 @@ void host_uart_close(int fd)
  * Name: host_uart_puts
  ****************************************************************************/
 
-ssize_t host_uart_puts(int fd, const char *buf, size_t size)
+int host_uart_puts(int fd, const char *buf, size_t size)
 {
   ssize_t ret;
 
@@ -151,7 +151,7 @@ ssize_t host_uart_puts(int fd, const char *buf, size_t size)
  * Name: host_uart_gets
  ****************************************************************************/
 
-ssize_t host_uart_gets(int fd, char *buf, size_t size)
+int host_uart_gets(int fd, char *buf, size_t size)
 {
   ssize_t ret;
