+++ arch/sim/src/sim/win/sim_hostuart.c
@@ -78,7 +78,7 @@ void host_uart_close(int fd)
  * Name: host_uart_putc
  ****************************************************************************/
 
-ssize_t host_uart_puts(int fd, const char *buf, size_t size)
+int host_uart_puts(int fd, const char *buf, size_t size)
 {
   DWORD nwritten;
   int ret;
@@ -92,7 +92,7 @@ ssize_t host_uart_puts(int fd, const char *buf, size_t size)
  * Name: host_uart_getc
  ****************************************************************************/
 
-ssize_t host_uart_gets(int fd, char *buf, size_t size)
+int host_uart_gets(int fd, char *buf, size_t size)
 {
   DWORD nread;
   int ret;
